third_party/blink/web_tests/platform/generic/fast/forms/form/htmlformelement-indexed-getter-expected.txt

This test verifies that elements can be fetched by index from HTMLFormElements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS form[0] is my_input
PASS successfullyParsed is true

TEST COMPLETE

HTMLInputElement
